appops: skip ops for invalid null package during state serialization
There's a bug that leads to construction of ops for invalid null package name.
Package name should always be non-null, it's defined and treated as such in AppOpsService.
It being null leads to crashes in system_server when appops state is serialized.

Previous commit reverted a buggy workaround for this bug, add a new workaround to prevent these
crashes.

Change-Id: I5262a347adebf73167e63a5314ca0941d8e39e2c